# docker-compose-lumen/src/storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Example API","contact":{"name":"Support Team","email":"support@example.com"},"version":"1.0"},"paths":{"\/src\/public\/auth\/register":{"post":{"tags":["Register User"],"operationId":"\/src\/public\/auth\/register","requestBody":{"description":"Pass user credentials","required":true,"content":{"application\/json":{"schema":{"required":["name","email","password"],"properties":{"name":{"type":"string","format":"string","example":"Ando"},"email":{"type":"string","format":"email","example":"user1@mail.com"},"password":{"type":"string","format":"string","example":"AndoAndo"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Authors","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}}}},"\/src\/public\/auth\/login":{"post":{"tags":["Login"],"operationId":"\/src\/public\/auth\/login","requestBody":{"description":"Pass user credentials","required":true,"content":{"application\/json":{"schema":{"required":["email","password"],"properties":{"email":{"type":"string","format":"eail","example":"user1@mail.com"},"password":{"type":"string","format":"string","example":"AndoAndo"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Authors","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}}}},"\/src\/public\/auth\/me":{"get":{"tags":["me"],"operationId":"\/src\/public\/auth\/me","responses":{"200":{"description":"Returns Authors","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}}}},"\/src\/public\/auth\/logout":{"post":{"tags":["logout"],"operationId":"\/src\/public\/auth\/logout","responses":{"200":{"description":"Returns Authors","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]}},"\/src\/public\/api\/categories":{"get":{"tags":["get all categories"],"operationId":"src\/public\/api\/categories","responses":{"200":{"description":"Returns Categorys","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]},"post":{"tags":["Create Category"],"operationId":"\/src\/public\/api\/categories","requestBody":{"description":"Pass Category name description","required":true,"content":{"application\/json":{"schema":{"required":["name","description"],"properties":{"name":{"type":"string","format":"string","example":"Category new"},"description":{"type":"string","format":"string","example":"discription"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Categories","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]}},"\/src\/public\/api\/categories\/{id}":{"get":{"tags":["Get Category"],"operationId":"\/src\/public\/api\/categories\/{id}","parameters":[{"name":"id","in":"path","description":"The Category id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns Categories","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Update Category"],"operationId":"\/src\/public\/api\/categories\/{id}","parameters":[{"name":"id","in":"path","description":"The Category id","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Pass Category name description","required":true,"content":{"application\/json":{"schema":{"required":["name","description"],"properties":{"name":{"type":"string","format":"string","example":"Category new"},"description":{"type":"string","format":"string","example":"description"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Categorys","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]},"delete":{"tags":["Delete Category"],"operationId":"\/src\/public\/api\/categories\/{id}","parameters":[{"name":"id","in":"path","description":"The Category id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns Categorys","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]}},"\/src\/public\/api\/products\/{categoryId}":{"post":{"tags":["Create Product"],"operationId":"\/src\/public\/api\/products\/{categoryId}","parameters":[{"name":"categoryId","in":"path","description":"The Category id","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Pass Product name description","required":true,"content":{"application\/json":{"schema":{"required":["name","description"],"properties":{"name":{"type":"string","format":"string","example":"Product new"},"description":{"type":"string","format":"string","example":"description"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Products","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]}},"\/src\/public\/api\/products\/{id}":{"put":{"tags":["Update Product"],"operationId":"\/src\/public\/api\/products\/{id}","parameters":[{"name":"id","in":"path","description":"The Product id","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Pass Category name description","required":true,"content":{"application\/json":{"schema":{"required":["name","description"],"properties":{"name":{"type":"string","format":"string","example":"Category new"},"description":{"type":"string","format":"string","example":"discription"}},"type":"object"}}}},"responses":{"200":{"description":"Returns Categorys","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]},"delete":{"tags":["Delete Product"],"operationId":"\/src\/public\/api\/products\/{id}","parameters":[{"name":"id","in":"path","description":"The Product id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns Products","content":{"application\/json":{"schema":{}}}},"400":{"description":"Error: Bad request. When required parameters were not supplied."}},"security":[{"bearer":[]}]}}},"components":{"securitySchemes":{"bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}